Clean methods and style category modal in create product
